// src/api/generated/model/updateIngredientBodyBaseUnit.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Bistro App Backend API
 * OpenAPI contract for generated frontend API clients.
 * OpenAPI spec version: 0.1.0
 */

export type UpdateIngredientBodyBaseUnit = typeof UpdateIngredientBodyBaseUnit[keyof typeof UpdateIngredientBodyBaseUnit];


export const UpdateIngredientBodyBaseUnit = {
  g: 'g',
  ml: 'ml',
} as const;